ml/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
numpy==1.26.4
pandas==1.5.0
scikit-learn==1.2.0
joblib==1.2.0
python-dotenv==1.0.0
requests==2.28.0
matplotlib==3.8.4
seaborn==0.13.2
shap==0.45.1
lime==0.2.0.1
gunicorn==21.2.0
kaggle==1.5.13
